pyproject.toml
[project]
name = "kahoot-clone"
version = "0.1.0"
description = ""
authors = [
    {name = "khanhtd",email = "khanhtd@ti-one.de"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "requests (>=2.32.5,<3.0.0)",
    "fastapi (>=0.118.0,<0.119.0)",
    "sqlalchemy[asyncio] (>=2.0.43,<3.0.0)",
    "argon2-cffi (>=25.1.0,<26.0.0)",
    "pymysql (>=1.1.2,<2.0.0)",
    "pydantic[email] (>=2.11.9,<3.0.0)",
    "pydantic-settings (>=2.11.0,<3.0.0)",
    "uvicorn (>=0.37.0,<0.38.0)",
    "aiomysql (>=0.2.0,<0.3.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "jwt (>=1.4.0,<2.0.0)",
    "redis (==5.0.1)",
]

[tool.poetry]
packages = [{include = "kahoot_clone", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
